Move search fixtures to ndbfixtures
