+++ nucliadb/tests/search/conftest.py
@@ -24,8 +24,8 @@ pytest_plugins = [
     "tests.ndbfixtures.processing",
     "tests.ndbfixtures.standalone",
     "tests.ingest.fixtures",  # should be refactored out
-    "tests.search.node",
-    "tests.search.fixtures",
+    "tests.ndbfixtures.node",
+    "tests.ndbfixtures.search",
     "nucliadb_utils.tests.fixtures",
     "nucliadb_utils.tests.gcs",
     "nucliadb_utils.tests.s3",
